refactor: move APIException in _exceptions module
