+++ hcloud/__init__.py
@@ -1,2 +1,2 @@
-from ._exceptions import HCloudException  # noqa
-from .hcloud import APIException, Client  # noqa
+from ._exceptions import APIException, HCloudException  # noqa
+from .hcloud import Client  # noqa
